api/src/services/sign-recognition/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.13.0
numpy==1.24.3
python-multipart==0.0.6
psycopg2-binary==2.9.7
pydantic==1.10.12
redis==4.5.5
celery==5.3.1
python-socketio==5.8.0
typing-extensions==4.5.0